server/requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.1.1
netifaces==0.10.4
oauthlib==2.0.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyserial==3.4
python-dateutil==2.8.1
python-debian==0.1.32
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.15.0
SQLAlchemy==1.3.20
systemd-python==234
Twisted==17.9.0
urllib3==1.22
Werkzeug==1.0.1
